delete previous controllers
