--- app/assets/javascripts/question.coffee
@@ -1,6 +1,0 @@
-# Place all the behaviors and hooks related to the matching controller here.
-# All this logic will automatically be available in application.js.
-# You can use CoffeeScript in this file: http://coffeescript.org/
-JQuery ->
-  $(document).ready ->
-    $('.new_answer').hide();
