--- app/helpers/answer_helper.rb
@@ -1,2 +1,0 @@
-module AnswerHelper
-end
