--- app/helpers/question_helper.rb
@@ -1,2 +1,0 @@
-module QuestionHelper
-end
