--- app/views/answer/index.html.haml
@@ -1,1 +1,0 @@
-%h2 index answer
