--- app/views/answer/new.html.haml
@@ -1,5 +1,0 @@
-= form_for @question.answers.new, url: question_answer_index_path(@question), method: :post do |a|
- = a.label :content, 'Form for Answer'
- = a.text_field :content
- = a.submit 'Provide an answer'
-%h3 new answer
