--- app/views/answer/show.html.haml
@@ -1,1 +1,0 @@
-%h2 answer show
