+++ app/views/answers/index.html.erb
@@ -1,2 +1,1 @@
-<h1>Answers#index</h1>
-<p>Find me in app/views/answers/index.html.erb</p>
+%h2 index answer
