+++ app/views/answers/show.html.erb
@@ -1,2 +1,1 @@
-<h1>Answers#show</h1>
-<p>Find me in app/views/answers/show.html.erb</p>
+%h2 answer show
