--- app/views/question/index.html.haml
@@ -1,6 +1,0 @@
-= link_to "Ask a question", new_question_path
-%br
-%h3 Question index
-- @questions.each do |question|
-    = link_to question.title, question_path(question)
-    %br
