+++ app/views/questions/index.html.erb
@@ -1,2 +1,6 @@
-<h1>Questions#index</h1>
-<p>Find me in app/views/questions/index.html.erb</p>
+= link_to "Ask a question", new_question_path
+%br
+%h3 Question index
+- @questions.each do |question|
+    = link_to question.title, question_path(question)
+    %br
